requirements.txt
pandas>=2.1.4
streamlit>=1.29.0
requests>=2.31.0
pillow>=10.2.0